Make static build honor "vite.resolve.*" config
In the clientBuild step, only some vite settings are passed to the final
vite command. "vite.resolve" was not included in this step, in spite of being
included in the ssrBuild step.

For example vite.resolve.alias, which is useful to use preact/compat
could not be configured in the astro.config file, as well as in
astro:config:setup hook.